Add Transaction entity and related services/repositories
This commit introduces a new `Transaction` entity along with its associated services, repositories, and controllers to manage transactions in the application. Key changes include:

- Implementation of `ITransactionService` and `ITransactionRepository` for transaction retrieval based on various criteria.
- Definition of the `Transaction` class with properties like `Title`, `Amount`, `TransactionDate`, `CategoryId`, `UserId`, and `Notes`.
- Updates to `ApplicationDbContext` to include a `DbSet<Transaction>` and configuration for relationships with `Category` and `User`.
- Creation of migration files to reflect the new `Transaction` table in the database schema.
- Modifications to `GenericRepository` and `IRepository` to use `long` for ID types for consistency.
- Addition of `TransactionController` to handle API requests related to transactions, with necessary service registrations in `Program.cs`.

+++ MoneyWiseLite.Domain/Interfaces/IRepository.cs
@@ -5,9 +5,9 @@ namespace MoneyWiseLite.Domain.Interfaces;
 public interface IRepository<T> where T : BaseEntity
 {
     Task<IEnumerable<T>> GetAllAsync();
-    Task<T?> GetByIdAsync(int id);
+    Task<T?> GetByIdAsync(long id);
     Task<T> AddAsync(T entity);
     Task UpdateAsync(T entity);
-    Task DeleteAsync(int id);
+    Task DeleteAsync(long id);
 }
